refactor(operator): extract conversation transitions
