+++ src/tui/components/operator-dashboard/index.tsx
@@ -96,6 +96,10 @@ import {
   tryParsePartialJson,
 } from "../shared";
 import {
+  recoverAbortedConversation,
+  rewriteToolResultOutput,
+} from "./conversation";
+import {
   appendStreamedText,
   applyToolCall,
   applyToolCallDelta,
@@ -108,7 +112,6 @@ import {
   buildOperatorSystemPrompt,
   type DashboardStatus,
   filterOperatorAutocomplete,
-  recoverAbortedConversation,
   resolveAbortAction,
   resolveInputFocused,
   resolveKeyboardShortcut,
@@ -1881,36 +1884,13 @@ This three-phase flow is specific to the TUI \`/threat-model\` command. The same
       const toolCallId = pendingToolCallIdRef.current;
       if (!toolCallId) return;
 
-      // Bedrock Converse / Anthropic requires { type: 'json', value: ... } wrapper.
-      const wrappedOutput = {
-        type: "json" as const,
-        value: result as unknown as Record<string, unknown>,
-      };
-
-      const updated: ModelMessage[] = conversationRef.current.map((msg) => {
-        if (msg.role !== "tool") return msg;
-        if (!Array.isArray(msg.content)) return msg;
-        let mutated = false;
-        const nextContent = msg.content.map((part) => {
-          if (
-            part &&
-            typeof part === "object" &&
-            "type" in part &&
-            (part as { type?: unknown }).type === "tool-result" &&
-            (part as { toolCallId?: unknown }).toolCallId === toolCallId
-          ) {
-            mutated = true;
-            return {
-              ...(part as Record<string, unknown>),
-              output: wrappedOutput,
-            };
-          }
-          return part;
-        });
-        return mutated
-          ? ({ ...msg, content: nextContent } as ModelMessage)
-          : msg;
-      });
+      // Bedrock Converse / Anthropic requires the { type: 'json', value }
+      // wrapper around the answers payload.
+      const updated = rewriteToolResultOutput(
+        conversationRef.current,
+        toolCallId,
+        result as unknown as Record<string, unknown>,
+      );
 
       conversationRef.current = updated;
 
@@ -2100,37 +2080,15 @@ This three-phase flow is specific to the TUI \`/threat-model\` command. The same
       key.preventDefault?.();
       const toolCallId = pendingToolCallIdRef.current;
       if (toolCallId) {
-        const abortedResult = {
-          type: "json" as const,
-          value: {
+        conversationRef.current = rewriteToolResultOutput(
+          conversationRef.current,
+          toolCallId,
+          {
             answers: [],
             skipped: true,
             aborted: true,
           } as unknown as Record<string, unknown>,
-        };
-        conversationRef.current = conversationRef.current.map((msg) => {
-          if (msg.role !== "tool" || !Array.isArray(msg.content)) return msg;
-          let mutated = false;
-          const nextContent = msg.content.map((part) => {
-            if (
-              part &&
-              typeof part === "object" &&
-              "type" in part &&
-              (part as { type?: unknown }).type === "tool-result" &&
-              (part as { toolCallId?: unknown }).toolCallId === toolCallId
-            ) {
-              mutated = true;
-              return {
-                ...(part as Record<string, unknown>),
-                output: abortedResult,
-              };
-            }
-            return part;
-          });
-          return mutated
-            ? ({ ...msg, content: nextContent } as ModelMessage)
-            : msg;
-        });
+        );
         const activeSession = sessionRef.current;
         if (activeSession) {
           try {
